Move `call_host_alloc` into `host` module
