+++ src/host/mod.rs
@@ -2,6 +2,8 @@
 
 pub mod api;
 
+use serde::Deserialize;
+
 use crate::module::{params_to_vec, Param, WasmModule};
 use crate::{LunaticError, ProcessConfig, Tag};
 
@@ -98,4 +100,28 @@ pub fn send_receive_skip_search(node: u64, process_id: u64, wait_on_tag: i64, ti
             api::distributed::send_receive_skip_search(node, process_id, wait_on_tag, timeout)
         }
     }
+}
+
+/// Utility for calling an allocating host function which is deserialized into
+/// `T`.
+///
+/// # Example
+///
+/// ```no_run
+/// struct Foo { a: String }
+///
+/// let foo = call_host_alloc::<Foo>(|len_ptr| unsafe {
+///     lunatic::host::some_allocating_fn(len_ptr)
+/// }).unwrap();
+/// ```
+#[doc(hidden)]
+pub fn call_host_alloc<T>(f: impl Fn(*mut u32) -> u32) -> bincode::Result<T>
+where
+    T: for<'de> Deserialize<'de>,
+{
+    let mut len = 0_u32;
+    let len_ptr = &mut len as *mut u32;
+    let ptr = f(len_ptr);
+    let data_vec = unsafe { Vec::from_raw_parts(ptr as *mut u8, len as usize, len as usize) };
+    bincode::deserialize(&data_vec)
 }
